# db_service/openapi.yaml
openapi: 3.0.1
info:
  description: ""
  license: {}
servers:
- url: /_db/test-workflows/torc-service
paths:
...
  /workflows:
    get:
      tags: []
      summary: Retrieve all workflows
      description: Retrieves all documents from the "workflows" collection.
      operationId: get_workflows
      parameters:
      - name: skip
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: number
          default: 0.0
      - name: limit
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: number
          default: 1000.0
      - name: name
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: user
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: description
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
      responses:
        "200":
          description: HTTP 200 OK.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_21'
        "500":
          description: Default error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_500'
      security: []
    post:
      tags: []
      summary: Store a workflow.
      description: Store a workflow in the "workflows" collection.
      operationId: post_workflows
      requestBody:
        description: Collection of jobs and dependent resources.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/workflows_model'
        required: true
      responses:
        "200":
          description: Collection of jobs and dependent resources
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/workflows_model'
        "500":
          description: Default error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_500'
      security: []
      x-codegen-request-body-name: body
  /workflows/{key}:
    get:
      tags: []
      summary: Retrieve the workflow for an key.
      description: Retrieve the document for a key from the "workflows" collection.
      operationId: get_workflows_key
      parameters:
      - name: key
        in: path
        description: key of the workflows document
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: HTTP 200 OK.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/workflows_model'
        "500":
          description: Default error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_500'
      security: []
    put:
      tags: []
      summary: Update workflow
      description: Update a document in the "workflows" collection.
      operationId: put_workflows_key
      parameters:
      - name: key
        in: path
        description: Key of the workflow.
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        description: workflow to update in the collection.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/workflows_key_model'
        required: true
      responses:
        "200":
          description: workflow updated in the collection.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/workflows_key_model'
        "500":
          description: Default error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_500'
      security: []
      x-codegen-request-body-name: body
    delete:
      tags: []
      summary: Delete a workflow
      description: Deletes a document from the "workflows" collection by key.
      operationId: delete_workflows_key
      parameters:
      - name: key
        in: path
        description: Workflow key.
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
        required: false
      responses:
        "200":
          description: workflow stored in the collection.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/workflows_key_model'
        "500":
          description: Default error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_500'
      security: []
      x-codegen-request-body-name: body
  /workflows/is_complete/{key}:
    get:
      tags: []
      summary: Report whether the workflow is complete
      description: Reports true if all jobs in the workflow are complete.
      operationId: get_workflows_is_complete_key
      parameters:
      - name: key
        in: path
        description: Workflow key
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: HTTP 200 OK.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_22'
        "500":
          description: Default error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_500'
      security: []
...
  /workflows/initialize_jobs/{key}:
    post:
      tags: []
      summary: Initialize job relationships.
      description: Initialize job relationships based on file relationships.
      operationId: post_workflows_initialize_jobs_key
      parameters:
      - name: key
        in: path
        description: Workflow key
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          '*/*': {}
        required: false
      responses:
        "200":
          description: message
          content:
            application/json:
              schema:
                type: object
        "500":
          description: Default error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_500'
      security: []
      x-codegen-request-body-name: body
  /workflows/prepare_jobs_for_submission/{key}:
    post:
      tags: []
      summary: Return ready jobs
      description: Return jobs that are ready for submission. Sets status to submitted_pending
      operationId: post_workflows_prepare_jobs_for_submission_key
      parameters:
      - name: key
        in: path
        description: Workflow key
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: limit
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: number
          default: 1000.0
      requestBody:
        description: Available worker resources.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/prepare_jobs_for_submission_key_model'
        required: true
      responses:
        "200":
          description: Jobs that are ready for submission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_24'
        "500":
          description: Default error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_500'
      security: []
      x-codegen-request-body-name: body
  /workflows/auto_tune_resource_requirements/{key}:
    post:
      tags: []
      summary: Enable workflow for auto-tuning resource requirements.
      description: Enable workflow for auto-tuning resource requirements.
      operationId: post_workflows_auto_tune_resource_requirements_key
      parameters:
      - name: key
        in: path
        description: Workflow key
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          '*/*': {}
        required: false
      responses:
        "200":
          description: Message
          content:
            application/json:
              schema:
                type: object
        "500":
          description: Default error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_500'
      security: []
      x-codegen-request-body-name: body
  /workflows/process_auto_tune_resource_requirements_results/{key}:
    post:
      tags: []
      summary: Process the results of auto-tuning resource requirements.
      description: Process the results of auto-tuning resource requirements.
      operationId: post_workflows_process_auto_tune_resource_requirements_results_key
      parameters:
      - name: key
        in: path
        description: Workflow key
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          '*/*': {}
        required: false
      responses:
        "200":
          description: Message
          content:
            application/json:
              schema:
                type: object
        "500":
          description: Default error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_500'
      security: []
      x-codegen-request-body-name: body
  /workflows/config/{key}:
    get:
      tags: []
      summary: Reports the workflow config.
      description: Reports the workflow config.
      operationId: get_workflows_config_key
      parameters:
      - name: key
        in: path
        description: Workflow key
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: HTTP 200 OK.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_18'
        "500":
          description: Default error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_500'
      security: []
    put:
      tags: []
      summary: Reports the workflow config.
      description: Reports the workflow config.
      operationId: put_workflows_config_key
      parameters:
      - name: key
        in: path
        description: Workflow key
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        description: Updated workflow config
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/config_key_model'
        required: true
      responses:
        "200":
          description: HTTP 200 OK.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_18'
        "500":
          description: Default error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_500'
      security: []
      x-codegen-request-body-name: body
  /workflows/cancel/{key}:
    put:
      tags: []
      summary: Cancel workflow.
      description: Cancel workflow. Workers will detect the status change and cancel
        jobs.
      operationId: put_workflows_cancel_key
      parameters:
      - name: key
        in: path
        description: Workflow key
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          '*/*': {}
        required: false
      responses:
        "200":
          description: message
          content:
            application/json:
              schema:
                type: object
        "500":
          description: Default error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_500'
      security: []
      x-codegen-request-body-name: body
  /workflows/reset_status/{key}:
    post:
      tags: []
      summary: Reset job status.
      description: Reset status for all jobs to uninitialized.
      operationId: post_workflows_reset_status_key
      parameters:
      - name: key
        in: path
        description: Workflow key
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          '*/*': {}
        required: false
      responses:
        "200":
          description: message
          content:
            application/json:
              schema:
                type: object
        "500":
          description: Default error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_500'
      security: []
      x-codegen-request-body-name: body
  /workflows/status/{key}:
    get:
      tags: []
      summary: Reports the workflow status.
      description: Reports the workflow status.
      operationId: get_workflows_status_key
      parameters:
      - name: key
        in: path
        description: Workflow key
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: HTTP 200 OK.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_25'
        "500":
          description: Default error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_500'
      security: []
    put:
      tags: []
      summary: Reports the workflow status.
      description: Reports the workflow status.
      operationId: put_workflows_status_key
      parameters:
      - name: key
        in: path
        description: Workflow key
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        description: Updated workflow status
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/status_key_model'
        required: true
      responses:
        "200":
          description: HTTP 200 OK.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_25'
        "500":
          description: Default error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_500'
      security: []
      x-codegen-request-body-name: body
  /workflows/collection_names/{key}:
    get:
      tags: []
      summary: Retrieve all collection names for one workflow.
      description: Retrieve all collection names for one workflow.
      operationId: get_workflows_collection_names_key
      parameters:
      - name: key
        in: path
        description: Workflow key
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: HTTP 200 OK.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_26'
        "500":
          description: Default error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_500'
      security: []
components:
  schemas:
...
    workflows_key_model:
      type: object
      properties:
        name:
          type: string
        user:
          type: string
        description:
          type: string
        _key:
          type: string
        _id:
          type: string
        _rev:
          type: string
    workflow_config_compute_node_resource_stats:
      type: object
      properties:
        cpu:
          type: boolean
          default: false
        disk:
          type: boolean
          default: false
        memory:
          type: boolean
          default: false
        network:
          type: boolean
          default: false
        process:
          type: boolean
          default: false
        include_child_processes:
          type: boolean
          default: true
        recurse_child_processes:
          type: boolean
          default: false
        monitor_type:
          type: string
          default: aggregation
        make_plots:
          type: boolean
          default: true
        interval:
          type: number
    inline_response_500:
      required:
      - error
      type: object
      properties:
        error:
          type: object
        errorNum:
          type: integer
        errorMessage:
          type: string
        code:
          type: integer
...
    inline_response_200_22:
      required:
      - is_canceled
      - is_complete
      type: object
      properties:
        is_canceled:
          type: boolean
        is_complete:
          type: boolean
    inline_response_200_21:
      required:
      - count
      - has_more
      - max_limit
      - skip
      - total_count
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/workflows_key_model'
        skip:
          type: integer
        max_limit:
          type: integer
        count:
          type: integer
        total_count:
          type: integer
        has_more:
          type: boolean
    inline_response_200_24:
      type: object
      properties:
        jobs:
          type: array
          items:
            $ref: '#/components/schemas/jobs_key_model'
        reason:
          type: string
    workflowsstatuskey_auto_tune_status:
      type: object
      properties:
        enabled:
          type: boolean
          default: true
        job_keys:
          type: array
          items:
            type: string
...
    inline_response_200_26:
      type: object
      properties:
        names:
          type: array
          items:
            type: string
    inline_response_200_25:
      required:
      - auto_tune_status
      - is_canceled
      - run_id
      type: object
      properties:
        is_canceled:
          type: boolean
        run_id:
          type: integer
        auto_tune_status:
          $ref: '#/components/schemas/workflowsstatuskey_auto_tune_status'
        _key:
          type: string
        _id:
          type: string
        _rev:
          type: string
...
    jobs_key_model:
      required:
      - command
      type: object
      properties:
        name:
          type: string
        command:
          type: string
        status:
          type: string
        cancel_on_blocking_job_failure:
          type: boolean
          default: true
        supports_termination:
          type: boolean
          default: false
        run_id:
          type: integer
        internal:
          $ref: '#/components/schemas/workflowsworkflowjobs_internal'
        _key:
          type: string
        _id:
          type: string
        _rev:
          type: string
...
    inline_response_200_18:
      type: object
      properties:
        compute_node_resource_stats:
          $ref: '#/components/schemas/workflow_config_compute_node_resource_stats'
        _key:
          type: string
        _id:
          type: string
        _rev:
          type: string
...
    status_key_model:
      required:
      - auto_tune_status
      - is_canceled
      - run_id
      type: object
      properties:
        is_canceled:
          type: boolean
        run_id:
          type: integer
        auto_tune_status:
          $ref: '#/components/schemas/workflowsstatuskey_auto_tune_status'
        _key:
          type: string
        _id:
          type: string
        _rev:
          type: string
...
    prepare_jobs_for_submission_key_model:
      required:
      - memory_gb
      - num_cpus
      type: object
      properties:
        num_cpus:
          type: integer
        memory_gb:
          type: number
        num_gpus:
          type: integer
        num_nodes:
          type: integer
        time_limit:
          type: object
        scheduler_config_id:
          type: string
    workflows_model:
      type: object
      properties:
        name:
          type: string
        user:
          type: string
        description:
          type: string
        _key:
          type: string
        _id:
          type: string
        _rev:
          type: string
...
    config_key_model:
      type: object
      properties:
        compute_node_resource_stats:
          $ref: '#/components/schemas/workflow_config_compute_node_resource_stats'
        _key:
          type: string
        _id:
          type: string
        _rev:
          type: string
...
    workflowsworkflowjobs_internal:
      type: object
      properties:
        memory_bytes:
          type: integer
        num_cpus:
          type: integer
        num_gpus:
          type: integer
        runtime_seconds:
          type: number
        scheduler_config_id:
          type: string
          default: ""
...